requirements.txt
kafka-python==2.0.2
psycopg2==2.8.6
unittest2==1.1.0
